Add files via upload
So, i mixed together the two programs for the GUI and I have an if statement that I need help with. 
I left a note in the code. Take a look at it and let me know if you guys can get it working or anything.
I'm not extremely keen on GUI commands but I think it's on the right track.